// fe/src/lib/api/generated/model/announcementStatusValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type AnnouncementStatusValue = typeof AnnouncementStatusValue[keyof typeof AnnouncementStatusValue];


export const AnnouncementStatusValue = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  ARCHIVED: 'ARCHIVED',
} as const;