// fe/src/lib/api/generated/model/announcementUpdateRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { AnnouncementAudienceValue } from './announcementAudienceValue';

export interface AnnouncementUpdateRequest {
  audience?: AnnouncementAudienceValue | null;
  body_markdown?: string | null;
  expires_at?: string | null;
  is_pinned?: boolean | null;
  title?: string | null;
}